fix(clippy): resolve all clippy warnings for Linux CI (-D warnings)
- Remove duplicate cfg(target_os = "linux") in platform/linux.rs
- Allow large_enum_variant on UserEvent (Config payload is intentionally inline)
- Allow too_many_arguments on mouse/scroll/header functions (coordinate-heavy APIs)
- Fix needless_range_loop in grid_renderer
- Remove unnecessary f32 cast on viewport scale_factor

+++ src/renderer/iced_layer.rs
@@ -314,7 +314,7 @@ impl IcedLayer {
     /// Run the iced UI lifecycle and present onto the given texture view.
     pub fn render(&mut self, view: &wgpu::TextureView, state: &UiState) -> Vec<UiMessage> {
         let bounds = self.viewport.logical_size();
-        let scale = self.viewport.scale_factor() as f32;
+        let scale = self.viewport.scale_factor();
 
         let widget = Self::view(state, scale);
